almost all tests

+++ routes/dashboard.js
@@ -13,40 +13,46 @@ function authenticateToken(req, res, next) {
     const authHeader = req.headers['authorization'];
 
     if (!authHeader) {
-        console.log(`Токена не хватает`);
-        return res.status(401).send('Токена не хватает');
+        console.error(`Отсутствует header "Authorization"`);
+        return res.status(401).send('Отсутствует header "Authorization"');
     }
 
     const token = authHeader.split(' ')[1];
 
     if (!token) {
-        console.log(`Токена не хватает`);
-        return res.status(401).send('Токена не хватает');
+        console.error(`Отсутствует токен`);
+        return res.status(401).send('Отсутствует токен');
     }
 
-    jwt.verify(token, process.env.TOKEN_SECRET, (err, data) => {
-      if (err) {
-        console.log('ERROR:', err.message);
-        return res.status(403).send(err.message);
-      }
-      next();
-    })
+    try {
+        jwt.verify(token, process.env.TOKEN_SECRET);
+        next();
+    } catch (error) {
+        console.error('ERROR:', error.message);
+        return res.status(403).send(error.message);
+    }
 }
 
 
 router.get('/dashboard', (req, res) => {
-    if (req.session.user) {
-        res.sendFile(path.join(__dirname, '../views/dashboard.html'));
-    } else {
-        res.redirect('/login');
-    }
+    res.sendFile(path.join(__dirname, '../views/dashboard.html'));
 });
 
 
 router.get('/tasks/todo', authenticateToken, async (req, res) => {
-    const user = req.session.user;
     try {
-        const tasks = await pool.query('SELECT t.* FROM tasks t LEFT JOIN user_to_task ut ON t.id = ut.task_id AND ut.user_id = $1 WHERE ut.task_id IS NULL;', [user.id]);
+        let userId;
+        try {
+            const user = req.session.user;
+            userId = user.id;
+        } catch {
+            const token = req.headers['authorization'].split(' ')[1];
+            const decoded = jwt.verify(token, process.env.TOKEN_SECRET);
+            const username = decoded.username;
+            const userQuery = await pool.query('SELECT * FROM users WHERE username = $1', [username]);
+            userId = userQuery.rows[0].id;
+        }
+        const tasks = await pool.query('SELECT t.* FROM tasks t LEFT JOIN user_to_task ut ON t.id = ut.task_id AND ut.user_id = $1 WHERE ut.task_id IS NULL;', [userId]);
         res.json(tasks.rows);
     } catch (err) {
         console.error(err);
@@ -55,9 +61,19 @@ router.get('/tasks/todo', authenticateToken, async (req, res) => {
 });
 
 router.get('/tasks/done', authenticateToken, async (req, res) => {
-    const user = req.session.user;
     try {
-        const user_tasks = await pool.query('SELECT t.*, ut.task_status FROM user_to_task ut JOIN users u ON ut.user_id = u.id JOIN tasks t ON ut.task_id = t.id where u.username=$1;', [user.username]);
+        let userId;
+        try {
+            const user = req.session.user;
+            userId = user.id;
+        } catch {
+            const token = req.headers['authorization'].split(' ')[1];
+            const decoded = jwt.verify(token, process.env.TOKEN_SECRET);
+            const username = decoded.username;
+            const userQuery = await pool.query('SELECT * FROM users WHERE username = $1', [username]);
+            userId = userQuery.rows[0].id;
+        }
+        const user_tasks = await pool.query('SELECT t.*, ut.task_status FROM user_to_task ut JOIN users u ON ut.user_id = u.id JOIN tasks t ON ut.task_id = t.id where u.id=$1;', [userId]);
         res.json(user_tasks.rows);
     } catch (err) {
         console.error(err);
@@ -66,10 +82,10 @@ router.get('/tasks/done', authenticateToken, async (req, res) => {
 });
 
 router.post('/tasks/:id/answer', authenticateToken, async (req, res) => {
-    const { answer } = req.body;
-    const { id } = req.params;
-    const user = req.session.user;
     try {
+        const { id } = req.params;
+        const user = req.session.user;
+        const { answer } = req.body;
         const taskAnswerQuery = await pool.query('SELECT answer FROM tasks WHERE id = $1', [id]);
         const taskAnswer = taskAnswerQuery.rows[0].answer;
         const taskStatus = taskAnswer == answer ? 1 : 0;
@@ -85,7 +101,18 @@ router.post('/tasks/:id/answer', authenticateToken, async (req, res) => {
 router.delete('/tasks/:id/undo', authenticateToken, async (req, res) => {
     const { id } = req.params;
     try {
-        await pool.query('DELETE FROM user_to_task WHERE task_id = $1', [id]);
+        let userId;
+        try {
+            const user = req.session.user;
+            userId = user.id;
+        } catch {
+            const token = req.headers['authorization'].split(' ')[1];
+            const decoded = jwt.verify(token, process.env.TOKEN_SECRET);
+            const username = decoded.username;
+            const userQuery = await pool.query('SELECT * FROM users WHERE username = $1', [username]);
+            userId = userQuery.rows[0].id;
+        }
+        await pool.query('DELETE FROM user_to_task WHERE task_id = $1 and user_id = $2', [id, userId]);
         res.status(204).send('Задача удалена из выполненных задач пользователя.');
     } catch (err) {
         console.error(err);
@@ -94,10 +121,3 @@ router.delete('/tasks/:id/undo', authenticateToken, async (req, res) => {
 });
 
 export default router;
-
-
-
-// status does not work with redirect. how to test it if even unauthorized returns 302?
-// tests for bad things, like unauthorized
-// delete user after registration test
-// tests coverage
